Moved PeakPost typings to shared
