+++ component-library/src/common/index.ts
@@ -2,6 +2,7 @@
  * @file Automatically generated by barrelsby.
  */
 
+export * from "./types";
 export * from "./image-loader/index";
 export * from "./media-embeds/index";
 export * from "./spinner/index";
